% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_TrainingData.R
\name{build_TrainingData}
\alias{build_TrainingData}
\title{Build training datasets for each species}
\usage{
build_TrainingData(
  GBIF_data,
  species,
  min.year,
  max.year,
  year.interval,
  save.unfiltered = FALSE
)
}
\arguments{
\item{GBIF_data}{Data frame containing occurrence records downloaded from GBIF.}

\item{species}{A character string or vector of character strings containing
the latin name of the species for which to build a training data set.}

\item{min.year}{Integer. Earliest year to include in training dataset.}

\item{max.year}{Integer. Latest year to include in training dataset.}

\item{year.interval}{Integer. Length of the time interval (number of years) for which 
to rasterize occurrence data.}

\item{save.unfiltered}{Logical, default = `FALSE`. If `TRUE`, unfiltered 
raster data on occurrence and sampling is saved in the subfolder Data/Raster.}
}
\value{
None
}
\description{
This function subsets, reformats, and rasterises occurrence and sampling data 
for a (list of) species, and saves the resulting training data sets in the
subfolder Data/Regression data.
}
