% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_IndicatorValues.R
\name{calculate_IndicatorValues}
\alias{calculate_IndicatorValues}
\title{Calculate and assemble new NI indicator values from GAM predictions}
\usage{
calculate_IndicatorValues(species, year, save = FALSE)
}
\arguments{
\item{species}{character string or vector of character strings containing
the latin name of the species for which to calculate indicator values}

\item{year}{vector of years for which to calculate indicator values}

\item{save}{logical, default = `FALSE`. If `TRUE`, new indicator data will be
saved in the working directory as `newIndicatorData.RData`.}
}
\value{
A list containing scaled indicator values and associated information
 for each municipality per `species` and `year`.
}
\description{
This function calculates scaled indicator values for each species- and year-
combination at the municipality level. 
The reference value used for scaling indicator values is extracted from
previously downloaded old indicator data (as stored in NI database) via the
label `Referanseverdi`.
}
