% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadData_GBIF.R
\name{downloadData_GBIF}
\alias{downloadData_GBIF}
\title{Download requested plant occurrence data for Norway from GBIF}
\usage{
downloadData_GBIF(key, path)
}
\arguments{
\item{key}{A character string representing a download key. Obtained via the
`getKey_GBIF` function, but can also be set manually if required.}

\item{path}{Directory into which GBIF data should be downloaded.}
}
\value{
A dataframe containing downloaded GBIF occurrence records.
}
\description{
This function downloads and unzips GBIF data that has been previously 
requested under a specified download key into a user-specified directory.
}
\details{
The procedure is based on the extensively commented instructions for
asynchronous downloading of GBIF data by Anders Finstad, available here:
https://gbif-europe.github.io/nordic_oikos_2018_r/s3_gbif_demo/3.x_async_download_gbif.Rmd
This allows downloading larger data sets, and citation of the download with 
a single doi. 

Note that after running `getKey_GBIF` to obtain the download key, you need 
to wait until GBIF has readied your dataset before you run `downloadData_GBIF`.
}
