% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadData_NIdb.R
\name{downloadData_NIdb}
\alias{downloadData_NIdb}
\title{Download indicator data from the Nature Index database}
\usage{
downloadData_NIdb(species, indicators, save = FALSE)
}
\arguments{
\item{species}{A vector of character strings representing Natin species names.}

\item{indicators}{A vector of character strings representing the Norwegian
names (without special characters) of the species listed in `species`. The
spelling of `indicators` is consistent with the spelling used in the NI 
database.}

\item{save}{Logical, default = `FALSE`. If `TRUE`, the downloaded indicator 
data is saved in the workspace in a file named `OldIndicatorData.RData`.}
}
\value{
A list of objects (one per `species`) containing 1) a data frame of 
the values of the values of the indicator and 2) a list of distribution 
objects quantifying the uncertainty of each indicator value estimate.
}
\description{
This function retrieves the currently stored values for one or several
indicators from the Nature Index (NI) database using the function 
`getIndicatorValues` from the `NICalc` package.
}
\details{
Executing this function requires accessing the NI database through a token
generated by `getToken`, and users are prompted to enter their NI database
log-in credentials for that purpose. NI database credentials consist of a
registered email address and a password.
}
\examples{

}
