% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_GAM.R
\name{predict_GAM}
\alias{predict_GAM}
\title{Make spatio-temporal occurrence predictions from GAM fits}
\usage{
predict_GAM(species, year, plot.pdf = TRUE, save = FALSE)
}
\arguments{
\item{species}{character string or vector of character strings containing
the latin name of the species for which to make predictions}

\item{year}{vector of years for which to make predictions}

\item{plot.pdf}{logical, default = `TRUE`. Whether or not to generate and
save pdf plots of predictions}

\item{save}{logical, default = `FALSE`. If `TRUE`, model fits for all 
species are saved in the subfolder Results as `NIGAM_All.list.RData`.}
}
\value{
A list containing for each `species` and `year` the mean and standard
deviation of predicted occurrence at the municipality level (stored as class
 `SpatialPolygonsDataFrame`).
}
\description{
This function uses the GAM fits to make occurrence predictions per 
municipality for each species-year combination.
}
\examples{

}
