% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_GAM.R
\name{run_GAM}
\alias{run_GAM}
\title{Run a generalized linear model (GAM) on spati-temporal occurrence data}
\usage{
run_GAM(species, save = TRUE)
}
\arguments{
\item{species}{A character string or vector of character strings containing
the latin name of the species for which to run the GAM}

\item{save}{Logical, default = `TRUE`. When `TRUE`, model fits for all 
species are saved in the subfolder Results as `gam.results.all.gamma3.RData`.
Set to `FALSE` to skip saving of new GAM fits.}
}
\value{
A list of GAM fits, one fit per species.
}
\description{
This function loads training data for a (list of) species and fits a 
space-time generalized additive model to it.
 
The response variable in `training_data` is the quantity `Y`, representing
the count of observations of a species within 1km by 1km cells over a 
five-year interval. The fitted effects are geographic coordinates (x and y
in meters), year (5-year intervals), and their two-way interactions, with a
smoothing parameter of 3.  
Sampling effort is included as an offset, and all data is analysed (no 
spatial precision filtering).
For more details on the modeling approach, consult earlier documentation
by O. Skarpaas (NI-2020-plants-report.docx).
}
\examples{

}
