% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadData_NIdb.R
\name{uploadData_NIdb}
\alias{uploadData_NIdb}
\title{Write updated indicator data to NI database}
\usage{
uploadData_NIdb(species, speciesList, mode)
}
\arguments{
\item{species}{character string or vector of character strings containing
the latin name of the species for which to upload/compare data}

\item{speciesList}{vector of character string containing latin names of all 
species in the set of indicators}

\item{mode}{switch for running different versions of the function. If 
`mode = "compare"` updated indicator data is compared to indicator data 
currently stored in the NI database. If `mode = "overwrite"`, updated 
indicator data is uploaded to the NI database.}
}
\value{

}
\description{
This function operates differently depending on the specified `mode` (see
Arguments).
}
\details{
In "compare" mode, the function downloads the current indicator data from the NI 
database and compares the newly assembled indicator with it. The comparison 
is displayed as proportional change in the new data relative to the data in 
the data base. 
In "overwrite" mode, the function uploads the newly assembled indicator data
to the NI database. Provided the user has writing access, this will overwrite
the data that is currently in the database and should ONLY be done when the 
database needs to be updated, and after the newly assembled indicator data
has been compared and quality-checked. 

When attempting to write to the NI database, the used will be prompted to 
confirm the action prior to its execution. 

Irrespective of the `mode`, executing this function requires accessing the NI
database through a token generated by `getToken`, and users are prompted to 
enter their NI database log-in credentials for that purpose. NI database 
credentials consist of a registered email address and a password. 


NOTE: As per now, unitOfMeasurement needs to be updated manually 
(new units = "Funnsannsynlighet i 1x1km-ruter")
}
